+++ tests/swig.test.ts
@@ -1,8 +1,8 @@
 import * as should from 'should';
-import { Swig } from '../lib/swig';
-import { statSync } from 'fs';
+import swig, { Swig } from '../lib/swig';
+import { statSync, readFileSync } from 'fs';
 
-describe('options', () => {
+describe('Options', () => {
     let swig: Swig;
     beforeEach(() => {
         swig = new Swig();
@@ -101,7 +101,7 @@ describe('options', () => {
     });
 })
 
-describe('separate instances', () => {
+describe('Separate instances', () => {
     it('properly autoescapes', () => {
         const a = new Swig({ autoescape: false }),
             b = new Swig();
@@ -110,31 +110,80 @@ describe('separate instances', () => {
     });
 });
 
-// describe('swog.compileFile', () => {
-//     const test = __dirname + '/cases/extend_1.test.html';
+describe('swog.compileFile', () => {
+    const test = __dirname + '/cases/extends_1.test.html';
 
-//     console.log(statSync(test));
+    let swig: Swig;
+    beforeEach(() => {
+        swig = new Swig();
+    });
+    afterEach(() => {
+        swig = null;
+    });
 
-//     let swig: Swig;
-//     beforeEach(() => {
-//         swig = new Swig();
-//     });
-//     afterEach(() => {
-//         swig = null;
-//     });
+    it('can run syncronously', () => {
+        should(swig.compileFile(test)()).be.ok();
+    });
 
-//     it('can run syncronously', () => {
-//         should(swig.compileFile(test)()).be.ok();
-//     });
+    it('can run asynchronously', (done) => {
+        // Run twice to ensure cached result uses the callback [gh-291]
+        swig.compileFile(test, {}, (err, fn) => {
+            should(fn).is.a.Function();
+            should(swig.compileFile(test, {}, (err, fn) => {
+                should(fn).is.a.Function();
+                done();
+            }));
+        });
+    });
 
-//     it('can run asynchronously', (done) => {
-//         // Run twice to ensure cached result uses the callback [gh-291]
-//         swig.compileFile(test, {}, (err, fn) => {
-//             should(fn).is.a.Function();
-//             should(swig.compileFile(test, {}, (err, fn) => {
-//                 should(fn).is.a.Function();
-//                 done();
-//             }));
-//         });
-//     });
-// })
+    it('can use callback with errors', function (done) {
+        let errorTest = __dirname + '/cases-error/extends-non-existent.test.html';
+        swig.compileFile(errorTest, {}, function (err) {
+            should(err.code).be.eql('ENOENT');
+            done();
+        });
+    });
+});
+
+
+describe('swig.renderFile', () => {
+    let test, expectation, s;
+
+    it('can use callback with errors occurred at the time of rendering', function (done) {
+        const s = new Swig({ loader: swig.loaders.memory({ 'error.html': '{{ foo() }}' }) });
+
+        s.renderFile('error.html', { foo: function () { throw new Error('bunk'); } }, function (err, out) {
+            should(err.message).be.eql('bunk');
+            done();
+        });
+    });
+
+    test = __dirname + '/cases/extends_1.test.html';
+    expectation = readFileSync(test.replace('test.html', 'expectation.html'), 'utf8');
+
+    beforeEach(() => {
+        s = new Swig();
+    });
+    afterEach(() => {
+        s = null;
+    });
+
+    it('can run syncronously', function () {
+        should(s.renderFile(test))
+            .be.eql(expectation);
+    });
+
+    it('can run asynchronously', function (done) {
+        s.renderFile(test, {}, function (err, fn) {
+            should(fn).be.eql(expectation);
+            done();
+        });
+    });
+
+    it('can use callbacks with errors', function (done) {
+        s.renderFile(__dirname + '/cases/not-existing', {}, function (err, out) {
+            should(err.code).be.eql('ENOENT');
+            done();
+        });
+    });
+})